Modulo-06-Funcionalidades/26-JWT/requirements.txt
ecdsa==0.18.0
pyasn1==0.4.8
PyJWT==2.5.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
